Major rework. Many things changed, see details ;)
- Add option for user-defined payload filter functions.
- Add License file (MIT)
- Add max_payload_length option and accumulated coverage output
- Enable tests to run as a suite. Not fully functional yet.
- Improving minimize
- Better module filter: no more absolute paths are required
- Improve Logging; add option to deactivate colors
- Add option to write a log file.
- Target function can now be given as symbol name
- Add support for UDP, multiple targets (loadbalancer). optimized coverage polling.
- Updated README.md with more detailed usage instructions

+++ tests/picohttpparser/run_tests.sh
@@ -6,29 +6,76 @@
 # [+] [seed=100]: speed=[59 exec/sec (avg: 52)] coverage=[175 bblocks] corpus=[31 files] last new path: [94] crashes: [0]
 # [+] [seed=1700]: speed=[35 exec/sec (avg: 41)] coverage=[188 bblocks] corpus=[42 files] last new path: [1639] crashes: [0]
 
-rm -rf tmpprojdir
+# Enable job control for shell script (so we can use 'fg', etc)
+set -m
 
-# old:
-# frizzer --project tmpprojdir --indir indir -f 0x401276 -t 7777 -p test -m /home/dennis/tools/frida-fuzzer/tests/picohttpparser/test
+exitfn () {
+    trap SIGINT
+    echo 'Interrupted by user!'
+    kill $test_pid
+    kill $frizzer_pid
+    exit
+}
+
+trap "exitfn" INT            # Set up SIGINT trap to call function.
+
+./test > /dev/null &
+test_pid=$!
+
+rm -rf tmpprojdir
 
 # new:
 frizzer init tmpprojdir
 cat > tmpprojdir/config <<EOF
 [fuzzer]
 log_level       = 3 # debug
+write_logfile   = true
 debug_mode      = false
+host            = "localhost"
+port            = 7777
+ssl             = false
+udp             = false
+fuzz_in_process = false
+recv_timeout    = 0.1
 
 [target]
 process_name    = "test"
 function        = 0x401276
-host            = "localhost"
-port            = 7777
 remote_frida    = false
-fuzz_in_process = false
+frida_port      = 27042
 modules = [
-        "/home/dennis/tools/frida-fuzzer/tests/picohttpparser/test",
+        "tests/picohttpparser/test",
     ]
 EOF
 
 frizzer add -p tmpprojdir indir
-frizzer fuzz -p tmpprojdir
+
+# start frizzer in the background
+frizzer fuzz -p tmpprojdir &
+frizzer_pid=$!
+
+# kill frizzer after 15 seconds
+(sleep 15; kill -s INT $frizzer_pid)&
+
+# get frizzer back in the foreground
+fg frizzer
+
+echo frizzer stopped!
+
+# check if fuzzer worked correctly:
+grep 'Found new path: \[41\] tmpprojdir/corpus/36_26_2_2' tmpprojdir/*.log
+result=$?
+
+echo "result: $result"
+
+if [ $result -eq 0 ]; then
+    echo "Test succeeded!"
+    rm -rf tmpprojdir
+else
+    echo "Test failed!"
+fi
+
+# cleanup
+kill $test_pid
+trap SIGINT
+exit $result
